"fix: アーキテクチャにそぐわないコードを修正"
